requirements.txt
autopep8==1.5.7
greenlet==1.1.0
mysql-connector-python==8.0.25
protobuf==3.17.3
pycodestyle==2.7.0
six==1.16.0
SQLAlchemy==1.4.19
toml==0.10.2
